Summary of changes:
* implement a much faster (and much more memory efficient) `ascii_image2`.
* Avoid building strings when lists/arrays will work.
* Avoid mutating list while iterating thru same list.
* Added an fps counter with `--fps` flag.

The result of this PR essentially doubles the framerate on my laptop (primarily because of the new `ascii_image2` implementation) from about 14fps to 30fps. It probably could go faster than 30fps, but my webcam only supports 30fps.