@groutr 2022-10-04
@joschuck I need to add that this has been very entertaining to play with!