@groutr 2022-10-05
The latest round of changes now gets processed at full framerate for my camera by resizing to target resolution before any processing. Preallocating arrays and reusing arrays also helps out.