@joschuck 2022-10-05
wow, I knew it was far for optimal but did not find the time, thanks for that! will have a look in a few hours and try to make a new release.